# Optimized encoding and decoding; simplified API