--- a/README.md
+++ b/README.md
@@ -1,7 +1,7 @@
 base62
 ------
 
-A simple base62 encoder/decoder, typically used in URL shorteners. It supports the [0-9A-Za-z] (Ordered) and [0-9a-zA-Z] (Reversed) variants.
+A simple base62 encoder/decoder, typically used in URL shorteners. It supports the standard [0-9A-Za-z] and alternative [0-9a-zA-Z] variants.
 
 [![Build status](https://github.com/fbernier/base62/workflows/ci/badge.svg)](https://github.com/fbernier/base62/actions)
 [![Crates.io](https://img.shields.io/crates/v/base62.svg)](https://crates.io/crates/base62)