# Improvements

* Encoding slightly faster and decoding much faster
* Encoding now appends to the given `String`, which is useful and which avoids the old behavior of just pointing the `String` reference at a completely new `String`, which is a usually-unnecessary allocation; if desired, the `String`'s contents can be replaced instead of appended to by `clear`ing the `String` before calling the encoding functions
* New `digit_count` function used to preallocate space in `String`s when encoding
* `Flavor` enum eliminated in favor of duplicated functions; this simplifies the API quite a bit
* Improved documentation, including examples that are true doctests (`assert_eq!` instead of `println!`)
* Added new `must_use`s and included the reason why the value must be used
* `DecodeError`'s `Debug` implementation now follows standard practice of producing a source code-like version of the error
* `DecodeError`'s invalid byte variant now stores a `u8` and now shows the `u8` in `b'_'` form when `Display` is used
* Renamed base62 variants to 'standard' and 'alternative'
* Simplified module structure by moving `base62` module to become the root of the crate

---

# Benchmarks

## Encoding with standard alphabet

### Old

    encode                  time:   [536.18 ns 537.69 ns 539.22 ns]
    Found 1 outliers among 100 measurements (1.00%)
      1 (1.00%) high mild

### New

    encode                  time:   [502.33 ns 503.82 ns 505.46 ns]
    Found 1 outliers among 100 measurements (1.00%)
      1 (1.00%) high severe
    
## Encoding with alternative alphabet

### Old

    encode_reversed         time:   [530.70 ns 532.18 ns 533.59 ns]
    Found 1 outliers among 100 measurements (1.00%)
      1 (1.00%) high mild

### New

    encode_alternative      time:   [499.17 ns 500.06 ns 501.01 ns]

## Decoding with standard alphabet

### Old

    decode                  time:   [340.28 ns 341.31 ns 342.40 ns]

### New

    decode                  time:   [59.846 ns 60.068 ns 60.290 ns]
    Found 2 outliers among 100 measurements (2.00%)
      1 (1.00%) low mild
      1 (1.00%) high severe

## Decoding with alternative alphabet

### Old

        decode_reversed         time:   [373.42 ns 376.42 ns 380.59 ns]
    Found 3 outliers among 100 measurements (3.00%)
      1 (1.00%) high mild
      2 (2.00%) high severe

### New

        decode_alternative      time:   [58.201 ns 58.353 ns 58.503 ns]
    Found 1 outliers among 100 measurements (1.00%)
      1 (1.00%) low mild