@ChaiTRex 2022-06-15
I only have a shared VPS, so benchmark times are a bit inexact for me. I have tried manually inlining and using `#[inline(…)]` hints and trying different refactorings and the newest commit is the best I was able to get.

    encode_standard_new     time:   [491.79 ns 493.08 ns 494.32 ns]
    Found 5 outliers among 100 measurements (5.00%)
      5 (5.00%) high mild
    
    encode_standard_buf     time:   [495.88 ns 496.85 ns 497.86 ns]
    
    encode_alternative_new  time:   [490.45 ns 491.59 ns 492.73 ns]
    
    encode_alternative_buf  time:   [498.27 ns 499.00 ns 499.78 ns]
    
    decode_standard         time:   [61.845 ns 62.681 ns 63.662 ns]                            
    Found 5 outliers among 100 measurements (5.00%)
      5 (5.00%) high severe
    
    decode_alternative      time:   [63.622 ns 63.769 ns 63.923 ns]
    Found 7 outliers among 100 measurements (7.00%)
      5 (5.00%) low mild
      1 (1.00%) high mild
      1 (1.00%) high severe

---

New changes:

* Long, repeated code snippets are moved to macros
* Refactoring made encoding functions about 10 to 15 ns faster
* `InvalidBase62Byte`'s `Display` implementation no longer allocates a `String`
* Added benchmarks for `encode_buf` and `encode_alternative_buf`
* Extended benchmark names so that none are prefixes of any others because `cargo bench prefix_of_benchmark_names` can be used to run only one benchmark, unless that benchmark's name is the prefix of another benchmark's name, which will cause it to run both