@ChaiTRex 2022-06-15
By using `u64` instead of `u128` division when possible, encoding is 3 times faster, decoding is slightly faster, and I assume but haven't measured that `digit_count` is faster.

    encode_standard_new     time:   [168.66 ns 169.25 ns 169.79 ns]
        
    encode_standard_buf     time:   [170.61 ns 171.21 ns 171.83 ns]
        
    encode_alternative_new  time:   [164.79 ns 165.56 ns 166.37 ns]
    Found 1 outliers among 100 measurements (1.00%)
      1 (1.00%) high mild
    
    encode_alternative_buf  time:   [171.31 ns 171.94 ns 172.60 ns]
    
    decode_standard         time:   [53.657 ns 53.839 ns 54.021 ns]
    Found 6 outliers among 100 measurements (6.00%)
      2 (2.00%) low mild
      1 (1.00%) high mild
      3 (3.00%) high severe
    
    decode_alternative      time:   [49.407 ns 49.853 ns 50.473 ns]
    Found 5 outliers among 100 measurements (5.00%)
      1 (1.00%) low mild
      1 (1.00%) high mild
      3 (3.00%) high severe