@ChaiTRex 2022-06-15
Small behavior change: decoding an empty input will give an error instead of 0 for a few reasons:

* Encoding will never give an empty output                                                                 
* Programmers might not realize that an empty input is an edge case that they might want to handle differently; this notifies them of the edge case
* With, for example, a URL-shortener, `https://what.ever/` should not have the same behavior as `https://what.ever/0`