@ChaiTRex 2022-06-15
Encoding is now slightly faster:

    encode_standard_new     time:   [155.10 ns 155.61 ns 156.17 ns]
    
    encode_standard_buf     time:   [160.26 ns 160.95 ns 161.65 ns]
    
    encode_alternative_new  time:   [156.22 ns 156.85 ns 157.44 ns]
    
    encode_alternative_buf  time:   [161.48 ns 163.52 ns 165.91 ns]
    Found 11 outliers among 100 measurements (11.00%)
      2 (2.00%) high mild
      9 (9.00%) high severe