@fbernier 2022-06-17
Published as 2.0.0

Thanks again!